fix(layering): move pure helpers to packages/utils out of business/db layers
Extracts broadcast, conversation, contact-filter-value-format, and
custom-field pure functions into packages/utils so action/query/API/job
layers stop importing across each other's boundaries.
