--- apps/builder/src/features/integration-messenger/message-templates/queries/index.ts
@@ -1,1 +1,0 @@
-export { messengerMessageTemplateService } from "@chatbotx.io/business"
