+++ packages/database/src/partials/conversation.ts
@@ -1,5 +1,17 @@
 import z from "zod"
 
+/**
+ * `inboxStatuses` is defined in `@chatbotx.io/utils/conversation` so a
+ * "use client" component (e.g. the broadcast inbox picker) can use it without
+ * depending on the database layer. Re-exported here because this has long
+ * been the import site for the rest of the repo; both paths resolve to the
+ * same enum. Mirrors the `channelTypes` precedent.
+ */
+export {
+  type InboxStatus,
+  inboxStatuses,
+} from "@chatbotx.io/utils/conversation"
+
 export const conversationBotCategories = z.enum(["bot", "human", "all"])
 export type ConversationBotCategory = z.infer<typeof conversationBotCategories>
 
@@ -15,9 +27,6 @@ export type ConversationStatus = z.infer<typeof conversationStatuses>
 export const assignerFilterTypes = z.enum(["all", "unassigned"])
 export type AssignerFilterType =
   (typeof assignerFilterTypes)[keyof typeof assignerFilterTypes]
-
-export const inboxStatuses = z.enum(["connected", "disconnected"])
-export type InboxStatus = z.infer<typeof inboxStatuses>
 
 export const inboxDisconnectReasons = z.enum([
   "manual",
